refactor(knowledge): one fixture harness, and an index that lists everything
The check/failures/passes harness had been copy-pasted byte-identically into
all four fixture suites -- four places to improve a failure message, and four
chances for one to drift into saying something slightly different about the
same kind of failure. It is one module now; the suites report identically and
their counts are unchanged (23 / 18 / 35 / 33).

The tools README's index listed twelve of the twenty-two modules. The six
missing ones included the whole changeset tool, named in no tracked file at
all, and api-surface/lib/surfaces.mjs, which knowledge/manifest.json points
readers at by path. Every non-test module is listed now, and the note about
how a fixture suite gets run says what the workflow actually does.
